backend/tests/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
Pillow==10.1.0

